Re-use compile button logic between embedded and full site
